// types.ts
export interface NgoData {
  name: string;
  focus: string;
  work: string;
  website: string;
}

export interface FactData {
  id: number;
  text: string;
  iconName: string;
}

export interface ImpactData {
  title: string;
  description: string;
}

export interface NavItem {
  label: string;
  href: string;
}